requirements.txt
uvicorn==0.20.0
fastapi
python-dotenv
sqlalchemy
sqlalchemy_serializer
asyncmy
cryptography